% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bucketize.cont}
\alias{bucketize.cont}
\title{Places a vector of continuous values into buckets of specified size.}
\usage{
bucketize.cont(vals, bucket.size = 1)
}
\arguments{
\item{vals}{A vector of numeric values that you which to place into buckets.}

\item{bucket.size}{Optional. Defaults to 1. The size bucket you wish to
create.}
}
\value{
A vector containing the bucketized values.
}
\description{
Places a vector of continuous values into buckets of specified size.
}
\examples{
vals = c(runif(n = 100, 1, 10), NA)
t = bucketize.cont(vals)
table(t)
v = bucketize.cont(vals, bucket.size = .5)
table(v)
}

