% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert.pct}
\alias{convert.pct}
\title{Convert a character string containing the percent sign , to an actual numeric percentage}
\usage{
convert.pct(pctStr)
}
\arguments{
\item{pctStr}{A string or vector of strings containing percentages.}
}
\value{
A numeric value or vector containing actual percentages.
}
\description{
Convert a character string containing the percent sign , to an actual numeric percentage
}
\examples{
convert.pct(c("10\%", "20.5\%", "foo", "bar", NA)) #Returns two percentages and 3 NA's
}

