% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.formula}
\alias{create.formula}
\title{Creates forumula from the given variables.}
\usage{
create.formula(y_var, x_vars, intercept = TRUE)
}
\arguments{
\item{y_var}{The variable you are interested in predicting.}

\item{x_vars}{The variables you believe may predict the y_var.}

\item{intercept}{Optional, defaults to TRUE. If FALSE, remove the intercept term from the formula.}
}
\value{
An R formula object.
}
\description{
Creates forumula from the given variables.
}
\examples{
x_vars <- c(as.character(paste0('X', c(1:10))), NA)
y_var <- 'Y'
create.formula(y_var, x_vars, intercept = FALSE)
}

