% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.mem.usage}
\alias{get.mem.usage}
\title{Displays the memory usage for all the objects in memory for R.}
\usage{
get.mem.usage(sort = "name", min.display.value = NA, limit = NA)
}
\arguments{
\item{sort}{Optional, defaults to name. If you select 'name', then it returns the objects in alphabetical order. If you select 'size', it returns the objects in descending order of size.}

\item{min.display.value}{Optional, defaults to NA. This parameter represents the smallest object that you want to display in your result set. For example, if you have hundreds of objects, but many of them are very small and inconsequential, you may want to select to min.display.value such that those objects are not shown.}

\item{limit}{Optional. Limit the resulting output to a specified size.}
}
\value{
A vector containing the objects and their associated sizes.
}
\description{
Displays the memory usage for all the objects in memory for R.
}
\examples{
get.mem.usage() # defaults to not filtering any values and sorting the results alphabetically.
get.mem.usage(sort = 'size')
get.mem.usage(sort = 'size', min.display.value = 10000)
get.mem.usage(limit = 10)
get.mem.usage(sort = 'size', min.display.value = 10000, limit = 10)
}

