% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpRfunctions-package.R
\docType{package}
\name{helpRfunctions-package}
\alias{helpRfunctions}
\alias{helpRfunctions-package}
\title{helpRfunctions: Collection of personal helper functions}
\description{
This package is a loose collection of my personal helper functions to effortlessly import them when I need to.
}
\author{
\strong{Maintainer}: Dennis Grothe \email{dennis.grothe@uni-bielefeld.de} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
