% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list.df.var.types}
\alias{list.df.var.types}
\title{Provides a list containing the data.frame columns that are of a given type. This can be useful for functions that only support a given type.}
\usage{
list.df.var.types(df)
}
\arguments{
\item{df}{The data.frame you wish to have a list of the types for.}
}
\value{
A list with names corresponding to the data type, and entries corresponding to the variable names that are of that type.
}
\description{
Provides a list containing the data.frame columns that are of a given type. This can be useful for functions that only support a given type.
}
\examples{
df <- data.frame(matrix(rnorm(100), nrow = 10))
df <- cbind(df, rep("A", nrow(df)))
cor(df) # This will return an error because the last column is not numeric.
varList <- list.df.var.types(df)
cor(df[,varList$numeric]) # No longer returns an error. Ignores the non-numeric columns.
}

