% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse.email}
\alias{parse.email}
\title{Splits an email address, or vector of email address, of the form
local-part@domain, into the local and domain pieces.}
\usage{
parse.email(emails)
}
\arguments{
\item{emails}{An email address, or vector of email addresses.}
}
\value{
A list containing the elements local and domain.
}
\description{
Splits an email address, or vector of email address, of the form
local-part@domain, into the local and domain pieces.
}
\examples{
emails = c('test323@gmail.com', 'not an email address', 'test2@email.org')
e = parse.email(emails)
e$local
e$domain
e = parse.email('test323@gmail.com')
e$local
e$domain
}

