% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.csv.dir}
\alias{read.csv.dir}
\title{Imports all the csv's in a given directory.}
\usage{
read.csv.dir(...)
}
\arguments{
\item{...}{Accepts all of the arguments for read.csv and list.files.}
}
\value{
All of the csv files imported
}
\description{
Imports all the csv's in a given directory.
}
\examples{
\dontrun{
data1 <- data.frame(1:10)
data2 <- data.frame(10:15)
write.csv(data1, file = 'data1.csv')
write.csv(data2, file = 'data2.csv')
read.csv.dir()
}
}

