% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rmv.highly.correlated}
\alias{rmv.highly.correlated}
\title{Removes highly correlated variables from your data frame. Ignores factor variables.}
\usage{
rmv.highly.correlated(df, verbose = FALSE, cutoff = 0.75,
  use = "pairwise.complete.obs", print.removed.cols = TRUE)
}
\arguments{
\item{df}{The incoming data frame}

\item{verbose}{Optional. Defaults to FALSE. Is the verbose argument for the \href{http://cran.r-project.org/web/packages/caret/caret.pdf}{findCorrelation} function in caret.}

\item{cutoff}{Optional. Defaults to .75. Is the cutoff argument for the \href{http://cran.r-project.org/web/packages/caret/caret.pdf}{findCorrelation} function in caret.}

\item{use}{Optional. Defaults to pairwise.complete.obs. Is the use argument for the cor function.}

\item{print.removed.cols}{Optional. Defaults to TRUE. If TRUE, prints the names of the columns removed.}
}
\value{
A new data frame with the highly correlated variables removed.
}
\description{
Removes highly correlated variables from your data frame. Ignores factor variables.
}
\examples{
df <- data.frame(matrix(rnorm(100), nrow = 10))
# Insert 10 random missing values
for (k in 1:10) {
 i = round(runif(n = 1, min = 0, max = 9))
 j = round(runif(n = 1, min = 0, max = 9))
 df[i, j] <- NA
}
df <- cbind(df, rep("A", nrow(df)))
df <- cbind(df, df[,1]) # TODO: This fails for this case! Name is null.
rmv.highly.correlated(df)
}

