% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rmv.one.level}
\alias{rmv.one.level}
\title{Removes variables that only contain one level.
Columns cotaning all NA's will also be removed.}
\usage{
rmv.one.level(df, print.removed.cols = TRUE)
}
\arguments{
\item{df}{The data.frame you wish to have a list of the types for.}

\item{print.removed.cols}{Optional, defaults to TRUE. If TRUE, prints the column names of the deleted columns.}
}
\value{
A data frame with variables only containing one level removed.
}
\description{
Removes variables that only contain one level.
Columns cotaning all NA's will also be removed.
}
\examples{
df <- data.frame(matrix(rnorm(100), nrow = 10))
for (k in 1:10) {
 i = round(runif(n = 1, min = 0, max = 9))
 j = round(runif(n = 1, min = 0, max = 9))
 df[i, j] <- NA
}
df <- cbind(df, rep("A", nrow(df)))
df <- cbind(df, df[,1])
df <- cbind(df, rep(NA, nrow(df)))
rmv.one.level(df)
}

