% Generated by roxygen2 (4.0.2): do not edit by hand
\name{top.factors}
\alias{top.factors}
\title{Return the top X factors and for the rest, group them together.}
\usage{
top.factors(df, factor.vars = names(df), max.factors = 10)
}
\arguments{
\item{df}{The incoming data frame}

\item{factor.vars}{Optional. Defaults to all columns in the df with names. The factor variables for which you want to group together.}

\item{max.factors}{Optional. Defaults to 10. The maximum number of factors you want to have for a given variable.}
}
\value{
A data frame with the factor.vars columns grouped by the max factors
}
\description{
Return the top X factors and for the rest, group them together.
}
\examples{
df <- data.frame(cbind(sample(letters, 100, replace = TRUE))
		,cbind(sample(letters, 100, replace = TRUE))
		,cbind(sample(letters, 100, replace = TRUE)))
df[1,1] <- NA
df <- cbind(df, df[,1])
levels(df[,ncol(df)]) <- c(levels(df[,ncol(df)]), 'another factor value', 'blah')
df[nrow(df),ncol(df)] <- 'another factor value'
df[nrow(df)-1,ncol(df)] <- 'blah'
names(df) <- 1:ncol(df)
top.factors(df)
top.factors(df, max.factors = 1000)
top.factors(df, max.factors = 27)
}

