% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trim}
\alias{trim}
\title{Trims whitespace from each element in a chacter vector.
Adapted from this post on \href{http://stackoverflow.com/questions/2261079/how-to-trim-leading-and-trailing-whitespace-in-r}{StackOverflow}}
\usage{
trim(x, white.space = "both")
}
\arguments{
\item{x}{A vector containing the strings to from which to remove white space.}

\item{white.space}{Optional. Acceptable values are "leading", "trailing", or "both". Defaults to both. Do you want to trim the "leading" white space, the "trailing" white space, or "both" the leading and trailing white space.}
}
\value{
A vector containing the bucketized values.
}
\description{
Trims whitespace from each element in a chacter vector.
Adapted from this post on \href{http://stackoverflow.com/questions/2261079/how-to-trim-leading-and-trailing-whitespace-in-r}{StackOverflow}
}
\examples{
x <-c(' leading whitespace', 'trailing whitespace ', ' both leading and trailing white space ')
trim(x)
trim(x, 'leading')
trim(x, 'trailing')
#trim(x, 'other') # throws error
}

