% Generated by roxygen2 (4.0.2): do not edit by hand
\name{split.data}
\alias{split.data}
\title{Divides the data into training and testing sets, with a specified percentage.
TODO: Extend this function to split into any number of groups with specified percentages in each group
TODO: Extend this function to work for matrices as well}
\usage{
\method{split}{data}(y, df, p = 0.7)
}
\arguments{
\item{y}{The column name of target variable, or variable of interest.}

\item{df}{The data frame to split}

\item{p}{Percent to put into the training data set. Optional, defaults to .7.}
}
\value{
A list containing data frames with training as the first element and the testing as the second
}
\description{
Divides the data into training and testing sets, with a specified percentage.
TODO: Extend this function to split into any number of groups with specified percentages in each group
TODO: Extend this function to work for matrices as well
}
\examples{
df <- data.frame(matrix(rnorm(100), nrow = 10))
t <- split.data(y = 'X1', df = df, p = .7)
training <- t[[1]]
testing <- t[[2]]
}

